+++ OpenNos.DAL.EF/Entities/Respawn.cs
@@ -24,7 +24,7 @@ namespace OpenNos.DAL.EF
 
         public virtual Map Map { get; set; }
 
-        public short MapId { get; set; }
+        public int MapId { get; set; }
 
         public long RespawnId { get; set; }
