+++ OpenNos.DAL.EF/Entities/RespawnMapType.cs
@@ -33,7 +33,7 @@ namespace OpenNos.DAL.EF
 
         #region Properties
 
-        public short DefaultMapId { get; set; }
+        public int DefaultMapId { get; set; }
 
         public short DefaultX { get; set; }
