+++ OpenNos.Data/TeleporterDTO.cs
@@ -20,7 +20,7 @@ namespace OpenNos.Data
 
         public short Index { get; set; }
 
-        public short MapId { get; set; }
+        public int MapId { get; set; }
 
         public int MapNpcId { get; set; }
