+++ OpenNos.Handler/CommandPacketHandler.cs
@@ -2013,7 +2013,7 @@ namespace OpenNos.Handler
                 {
                     return;
                 }
-                if (short.TryParse(teleportPacket.Data, out short mapId))
+                if (int.TryParse(teleportPacket.Data, out int mapId))
                 {
                     ServerManager.Instance.ChangeMap(Session.Character.CharacterId, mapId, teleportPacket.X, teleportPacket.Y);
                 }
